new gases code cleanup and bug fixes
