+++ Content.Server/_Funkystation/Atmos/Reactions/HyperNobliumProductionReaction.cs
@@ -1,9 +1,10 @@
+using Content.Server.Atmos;
 using Content.Server.Atmos.EntitySystems;
 using Content.Shared.Atmos;
 using Content.Shared.Atmos.Reactions;
 using JetBrains.Annotations;
 
-namespace Content.Server.Atmos.Reactions;
+namespace Content.Server._Funkystation.Atmos.Reactions;
 
 [UsedImplicitly]
 public sealed partial class HyperNobliumProductionReaction : IGasReactionEffect
@@ -12,7 +13,7 @@ public sealed partial class HyperNobliumProductionReaction : IGasReactionEffect
     {
         if (mixture.Temperature > 20f && mixture.GetMoles(Gas.HyperNoblium) >= 5f)
             return ReactionResult.NoReaction;
-            
+
         var initN2 = mixture.GetMoles(Gas.Nitrogen);
         var initTrit = mixture.GetMoles(Gas.Tritium);
         var initBZ = mixture.GetMoles(Gas.BZ);
@@ -21,14 +22,14 @@ public sealed partial class HyperNobliumProductionReaction : IGasReactionEffect
         var reductionFactor = Math.Clamp(initTrit / (initTrit + initBZ), 0.001f, 1f);
 
         var nobFormed = Math.Min((initN2 + initTrit) * 0.01f, Math.Min(initTrit * 1f / (5f * reductionFactor), initN2 * 0.1f));
-        if (nobFormed <= 0 || (initTrit - 5f) * nobFormed * reductionFactor < 0 || (initN2 - 10f) * nobFormed < 0)
+        if (nobFormed <= 0f || initTrit - 5f * nobFormed * reductionFactor < 0f || initN2 - 10f * nobFormed < 0f)
             return ReactionResult.NoReaction;
 
         mixture.AdjustMoles(Gas.Tritium, -5f * nobFormed * reductionFactor);
         mixture.AdjustMoles(Gas.Nitrogen, -10f * nobFormed);
         mixture.AdjustMoles(Gas.HyperNoblium, nobFormed);
 
-        var energyReleased = nobFormed * (Atmospherics.HyperNobliumProductionEnergy/Math.Max(initBZ, 1));
+        var energyReleased = nobFormed * (Atmospherics.HyperNobliumProductionEnergy / Math.Max(initBZ, 1f));
 
         var heatCap = atmosphereSystem.GetHeatCapacity(mixture, true);
         if (heatCap > Atmospherics.MinimumHeatCapacity)
