+++ Content.Server/_Funkystation/Atmos/Reactions/ProtoNitrateBZConversionReaction.cs
@@ -1,9 +1,10 @@
+using Content.Server.Atmos;
 using Content.Server.Atmos.EntitySystems;
 using Content.Shared.Atmos;
 using Content.Shared.Atmos.Reactions;
 using JetBrains.Annotations;
 
-namespace Content.Server.Atmos.Reactions;
+namespace Content.Server._Funkystation.Atmos.Reactions;
 
 [UsedImplicitly]
 public sealed partial class ProtoNitrateBZConversionReaction : IGasReactionEffect
