+++ Content.Server/_Funkystation/Atmos/Reactions/ZaukerDecompositionReaction.cs
@@ -1,30 +1,32 @@
+using Content.Server.Atmos;
 using Content.Server.Atmos.EntitySystems;
 using Content.Shared.Atmos;
 using Content.Shared.Atmos.Reactions;
 using JetBrains.Annotations;
 
-namespace Content.Server.Atmos.Reactions;
+namespace Content.Server._Funkystation.Atmos.Reactions;
 
 [UsedImplicitly]
-public sealed partial class ProtoNitrateHydrogenConversionReaction : IGasReactionEffect
+public sealed partial class ZaukerDecompositionReaction : IGasReactionEffect
 {
     public ReactionResult React(GasMixture mixture, IGasMixtureHolder? holder, AtmosphereSystem atmosphereSystem, float heatScale)
     {
         if (mixture.Temperature > 20f && mixture.GetMoles(Gas.HyperNoblium) >= 5f)
             return ReactionResult.NoReaction;
 
-        var initPN = mixture.GetMoles(Gas.ProtoNitrate);
-        var initH2 = mixture.GetMoles(Gas.Hydrogen);
+        var initZauker = mixture.GetMoles(Gas.Zauker);
+        var initN2 = mixture.GetMoles(Gas.Nitrogen);
 
-        var producedAmount = Math.Min(Atmospherics.ProtoNitrateHydrogenConversionMaxRate, Math.Min(initH2, initPN));
+        var burnedFuel = Math.Min(Atmospherics.ZaukerDecompositionMaxRate, Math.Min(initN2, initZauker));
 
-        if (producedAmount <= 0 || initH2 - producedAmount < 0f)
+        if (burnedFuel <= 0 || initZauker - burnedFuel < 0)
             return ReactionResult.NoReaction;
 
-        mixture.AdjustMoles(Gas.Hydrogen, -producedAmount);
-        mixture.AdjustMoles(Gas.ProtoNitrate, producedAmount * 0.5f);
+        mixture.AdjustMoles(Gas.Zauker, -burnedFuel);
+        mixture.AdjustMoles(Gas.Oxygen, burnedFuel * 0.3f);
+        mixture.AdjustMoles(Gas.Nitrogen, burnedFuel * 0.7f);
 
-        var energyReleased = producedAmount * Atmospherics.ProtoNitrateHydrogenConversionEnergy;
+        var energyReleased = burnedFuel * Atmospherics.ZaukerDecompositionEnergy;
 
         var heatCap = atmosphereSystem.GetHeatCapacity(mixture, true);
         if (heatCap > Atmospherics.MinimumHeatCapacity)
